# Commit Message
## Description
Needed to make UI more intuitive. By moving the save and add column button to the end, users will be able to read their inputs on the Table Name field.

## Issues
After making the changes, the Save button on TABLE_FOOTER is currently not functioning as expected.

## Changes Made
Added TABLE_FOOTER to display after TABLE_HEADER and TABLE_FIELD. Removed unnecessary code that was not functional.

# Co-authored-by

Co-authored-by: Elaine Wong <ewong1217@gmail.com>
Co-authored-by:  Cathy Luong<cathyluong93@yahoo.com>
Co-authored-by: Zoren Labrador <zorenzal@gmail.com>